Add namespace support and attribute support

+++ misc/xml-to-record-converter/src/main/java/io/ballerina/xmltorecordconverter/util/ConverterUtils.java
@@ -1,7 +1,7 @@
 /*
- *  Copyright (c) 2022, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *  Copyright (c) 2023, WSO2 LLC. (http://www.wso2.org)
  *
- *  WSO2 Inc. licenses this file to you under the Apache License,
+ *  WSO2 LLC. licenses this file to you under the Apache License,
  *  Version 2.0 (the "License"); you may not use this file except
  *  in compliance with the License.
  *  You may obtain a copy of the License at
@@ -39,7 +39,7 @@ import static io.ballerina.identifier.Utils.unescapeUnicodeCodepoints;
 /**
  * Util methods for XML to record converter.
  *
- * @since 2201.2.0
+ * @since 2201.7.2
  */
 public class ConverterUtils {
 
@@ -163,7 +163,7 @@ public class ConverterUtils {
 
     private static boolean isInteger(String value) {
         try {
-            Integer.parseInt(value);
+            Long.parseLong(value);
             return true;
         } catch (NumberFormatException e) {
             return false;
